+++ federation/pkg/kubefed/join_test.go
@@ -19,27 +19,23 @@ package kubefed
 import (
 	"bytes"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"net/http"
-	"os"
 	"testing"
 
 	federationapi "k8s.io/kubernetes/federation/apis/federation/v1beta1"
+	kubefedtesting "k8s.io/kubernetes/federation/pkg/kubefed/testing"
 	"k8s.io/kubernetes/federation/pkg/kubefed/util"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/testapi"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/api/v1"
-	"k8s.io/kubernetes/pkg/apimachinery/registered"
-	"k8s.io/kubernetes/pkg/client/restclient"
 	"k8s.io/kubernetes/pkg/client/restclient/fake"
 	"k8s.io/kubernetes/pkg/client/typed/dynamic"
 	"k8s.io/kubernetes/pkg/client/unversioned/clientcmd"
 	clientcmdapi "k8s.io/kubernetes/pkg/client/unversioned/clientcmd/api"
 	cmdtesting "k8s.io/kubernetes/pkg/kubectl/cmd/testing"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
-	"k8s.io/kubernetes/pkg/runtime"
 )
 
 func TestJoinFederation(t *testing.T) {
@@ -48,11 +44,11 @@ func TestJoinFederation(t *testing.T) {
 		cmdErrMsg = str
 	})
 
-	fakeKubeFiles, err := fakeKubeconfigFiles()
+	fakeKubeFiles, err := kubefedtesting.FakeKubeconfigFiles()
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
-	defer rmFakeKubeconfigFiles(fakeKubeFiles)
+	defer kubefedtesting.RemoveFakeKubeconfigFiles(fakeKubeFiles)
 
 	testCases := []struct {
 		cluster            string
@@ -111,7 +107,7 @@ func TestJoinFederation(t *testing.T) {
 			t.Fatalf("[%d] unexpected error: %v", i, err)
 		}
 
-		adminConfig, err := newFakeAdminConfig(hostFactory, tc.kubeconfigGlobal)
+		adminConfig, err := kubefedtesting.NewFakeAdminConfig(hostFactory, tc.kubeconfigGlobal)
 		if err != nil {
 			t.Fatalf("[%d] unexpected error: %v", i, err)
 		}
@@ -162,7 +158,7 @@ func testJoinFederationFactory(name, server string) cmdutil.Factory {
 				if !api.Semantic.DeepEqual(got, want) {
 					return nil, fmt.Errorf("unexpected cluster object\n\tgot: %#v\n\twant: %#v", got, want)
 				}
-				return &http.Response{StatusCode: http.StatusCreated, Header: defaultHeader(), Body: objBody(codec, &want)}, nil
+				return &http.Response{StatusCode: http.StatusCreated, Header: kubefedtesting.DefaultHeader(), Body: kubefedtesting.ObjBody(codec, &want)}, nil
 			default:
 				return nil, fmt.Errorf("unexpected request: %#v\n%#v", req.URL, req)
 			}
@@ -170,30 +166,6 @@ func testJoinFederationFactory(name, server string) cmdutil.Factory {
 	}
 	tf.Namespace = "test"
 	return f
-}
-
-type fakeAdminConfig struct {
-	pathOptions *clientcmd.PathOptions
-	hostFactory cmdutil.Factory
-}
-
-func newFakeAdminConfig(f cmdutil.Factory, kubeconfigGlobal string) (util.AdminConfig, error) {
-	pathOptions := clientcmd.NewDefaultPathOptions()
-	pathOptions.GlobalFile = kubeconfigGlobal
-	pathOptions.EnvVar = ""
-
-	return &fakeAdminConfig{
-		pathOptions: pathOptions,
-		hostFactory: f,
-	}, nil
-}
-
-func (f *fakeAdminConfig) PathOptions() *clientcmd.PathOptions {
-	return f.pathOptions
-}
-
-func (f *fakeAdminConfig) HostFactory(host, kubeconfigPath string) cmdutil.Factory {
-	return f.hostFactory
 }
 
 func fakeJoinHostFactory(name, server, token string) (cmdutil.Factory, error) {
@@ -227,7 +199,7 @@ func fakeJoinHostFactory(name, server, token string) (cmdutil.Factory, error) {
 		},
 		ObjectMeta: v1.ObjectMeta{
 			Name:      name,
-			Namespace: "federation-system",
+			Namespace: util.DefaultFederationSystemNamespace,
 		},
 		Data: map[string][]byte{
 			"kubeconfig": configBytes,
@@ -236,7 +208,7 @@ func fakeJoinHostFactory(name, server, token string) (cmdutil.Factory, error) {
 
 	f, tf, codec, _ := cmdtesting.NewAPIFactory()
 	ns := dynamic.ContentConfig().NegotiatedSerializer
-	tf.ClientConfig = defaultClientConfig()
+	tf.ClientConfig = kubefedtesting.DefaultClientConfig()
 	tf.Client = &fake.RESTClient{
 		NegotiatedSerializer: ns,
 		Client: fake.CreateHTTPClient(func(req *http.Request) (*http.Response, error) {
@@ -254,7 +226,7 @@ func fakeJoinHostFactory(name, server, token string) (cmdutil.Factory, error) {
 				if !api.Semantic.DeepEqual(got, secretObject) {
 					return nil, fmt.Errorf("Unexpected secret object\n\tgot: %#v\n\twant: %#v", got, secretObject)
 				}
-				return &http.Response{StatusCode: http.StatusCreated, Header: defaultHeader(), Body: objBody(codec, &secretObject)}, nil
+				return &http.Response{StatusCode: http.StatusCreated, Header: kubefedtesting.DefaultHeader(), Body: kubefedtesting.ObjBody(codec, &secretObject)}, nil
 			default:
 				return nil, fmt.Errorf("unexpected request: %#v\n%#v", req.URL, req)
 			}
@@ -278,116 +250,6 @@ func fakeCluster(name, server string) federationapi.Cluster {
 			SecretRef: &v1.LocalObjectReference{
 				Name: name,
 			},
-		},
-	}
-}
-
-func fakeKubeconfigFiles() ([]string, error) {
-	kubeconfigs := []clientcmdapi.Config{
-		{
-			Clusters: map[string]*clientcmdapi.Cluster{
-				"syndicate": {
-					Server: "https://10.20.30.40",
-				},
-			},
-			AuthInfos: map[string]*clientcmdapi.AuthInfo{
-				"syndicate": {
-					Token: "badge",
-				},
-			},
-			Contexts: map[string]*clientcmdapi.Context{
-				"syndicate": {
-					Cluster:  "syndicate",
-					AuthInfo: "syndicate",
-				},
-			},
-			CurrentContext: "syndicate",
-		},
-		{
-			Clusters: map[string]*clientcmdapi.Cluster{
-				"ally": {
-					Server: "ally256.example.com:80",
-				},
-			},
-			AuthInfos: map[string]*clientcmdapi.AuthInfo{
-				"ally": {
-					Token: "souvenir",
-				},
-			},
-			Contexts: map[string]*clientcmdapi.Context{
-				"ally": {
-					Cluster:  "ally",
-					AuthInfo: "ally",
-				},
-			},
-			CurrentContext: "ally",
-		},
-		{
-			Clusters: map[string]*clientcmdapi.Cluster{
-				"ally": {
-					Server: "https://ally64.example.com",
-				},
-				"confederate": {
-					Server: "10.8.8.8",
-				},
-			},
-			AuthInfos: map[string]*clientcmdapi.AuthInfo{
-				"ally": {
-					Token: "souvenir",
-				},
-				"confederate": {
-					Token: "totem",
-				},
-			},
-			Contexts: map[string]*clientcmdapi.Context{
-				"ally": {
-					Cluster:  "ally",
-					AuthInfo: "ally",
-				},
-				"confederate": {
-					Cluster:  "confederate",
-					AuthInfo: "confederate",
-				},
-			},
-			CurrentContext: "confederate",
-		},
-	}
-	kubefiles := []string{}
-	for _, cfg := range kubeconfigs {
-		fakeKubeFile, _ := ioutil.TempFile("", "")
-		err := clientcmd.WriteToFile(cfg, fakeKubeFile.Name())
-		if err != nil {
-			return nil, err
-		}
-
-		kubefiles = append(kubefiles, fakeKubeFile.Name())
-	}
-	return kubefiles, nil
-}
-
-func rmFakeKubeconfigFiles(kubefiles []string) {
-	for _, file := range kubefiles {
-		os.Remove(file)
-	}
-}
-
-func defaultHeader() http.Header {
-	header := http.Header{}
-	header.Set("Content-Type", runtime.ContentTypeJSON)
-	return header
-}
-
-func objBody(codec runtime.Codec, obj runtime.Object) io.ReadCloser {
-	return ioutil.NopCloser(bytes.NewReader([]byte(runtime.EncodeOrDie(codec, obj))))
-}
-
-func defaultClientConfig() *restclient.Config {
-	return &restclient.Config{
-		APIPath: "/api",
-		ContentConfig: restclient.ContentConfig{
-			NegotiatedSerializer: api.Codecs,
-			ContentType:          runtime.ContentTypeJSON,
-			GroupVersion:         &registered.GroupOrDie(api.GroupName).GroupVersion,
 		},
 	}
 }
